refactor: add file with common types and interface
