+++ webpack.config.js
@@ -4,7 +4,7 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 
 const baseConfig = {
-	entry: path.resolve(__dirname, './src/index.js'),
+	entry: path.resolve(__dirname, './src/index.ts'),
 	mode: 'development',
 	module: {
 		rules:
